+++ src/main/resources/templates/myPage/info.html
@@ -32,6 +32,7 @@
                                 <div>내 정보 조회 및 수정 병합은 여기서 가능해요</div>
                             </div>
                         </div>
+                        <hr>
                         <div>
                             <div class="row  my-3">
                                 <div class="col">
